requirements.txt
jupyter
pandas
pyyaml
nbstripout
matplotlib
numpy
scipy
tqdm
flake8